App이 Todo의 상태를 관리하게하라
